cmc api, kraken api update

+++ src/main/java/cz/tomastokamrazek/arion/datapoint/kraken/dto/KrakenLedger.java
@@ -16,7 +16,7 @@ public class KrakenLedger {
 	String refId;
 	
 	@JsonProperty("time")
-	Double time;
+	BigDecimal time;
 	
 	@JsonProperty("type")
 	KrakenLedgerType type;
